refactor(UserException) : 사용자 예외 도메인 패키지 분리
기존 `domain` 패키지에 있던 `UserException`과 관련 클래스를 `domain.user` 패키지로 이동하여 도메인별 책임을 명확히 분리하였습니다.
- `UserException`, `UserErrorCode`의 위치 변경
- 관련된 import 문 수정
