+++ src/main/kotlin/com/terning/server/kotlin/domain/user/UserErrorCode.kt
@@ -1,4 +1,4 @@
-package com.terning.server.kotlin.domain
+package com.terning.server.kotlin.domain.user
 
 import org.springframework.http.HttpStatus
